+++ curious/core/state.py
@@ -370,7 +370,6 @@ class State(object):
 
         author_id = int(event_data.get("author", {}).get("id", 0))
 
-        message.channel = channel
         if channel is not None:
             message.guild_id = channel.guild_id
         if message.channel.type == ChannelType.PRIVATE: